Pig_Latin_Problem
Pig Latin - Pig Latin is a game of alterations played in the English language game. To create the Pig Latin form of an English word the initial consonant sound is transposed to the end of the word and an ay is affixed (Ex.: "banana" would yield anana-bay). Read Wikipedia for more information on rules.

Link: https://en.wikipedia.org/wiki/Pig_Latin#:~:text=Pig%20Latin%2C%20or%20%22Igpay%20Atinlay,to%20create%20such%20a%20suffix.